+++ src/moai-core/MOAISerializer.cpp
@@ -611,7 +611,7 @@ u32 MOAISerializer::WriteTableInitializer ( ZLStream& stream, MOAILuaState& stat
 			}
 			case LUA_TUSERDATA: {
 				MOAILuaObject* object = state.GetLuaObject < MOAILuaObject >( -1, false );
-				u32 instanceID = this->GetID ( object );
+				uintptr instanceID = this->GetID ( object );
 				stream.Print ( "objects [ 0x%08X ]\n", instanceID );
 				break;
 			}
